backend/requirements.txt
requests==2.26.0
django==2.2.16
djangorestframework==3.12.4
gunicorn==20.0.4
PyJWT==2.1.0
django-filter==21.1
psycopg2-binary==2.8.6
python-dotenv==0.20.0
djoser
Pillow==9.2.0
webcolors==1.12